Added Jupyter notebook for single Cp calculation.
A Jupyter notebook with a step-by-step calculation of the Cp. It also contains a simple visualization of the probability density. The mean, the standard deviation, and the specification limits can be changed as desired. The Cp index, the probability for a value outside the specification limits, and PPM are calculated.